# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: QuoteMyAV API
  description: |
    REST API for QuoteMyAV - AI-powered AV quote generation platform.

    ## Authentication

    The API supports two authentication methods:

    ### JWT Authentication (Web Users)
    Use Supabase JWT tokens in the `Authorization` header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ### API Key Authentication (Pro/Enterprise)
    Use API keys in the `X-API-Key` header:
    ```
    X-API-Key: qmav_live_xxxxx
    ```

    ## Rate Limits

    | Tier | API Calls/Day | AI Calls/Hour |
    |------|---------------|---------------|
    | Pro | 100 | 20 |
    | Enterprise | 10,000 | 500 |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Quotas

    | Tier | Monthly Quote Limit |
    |------|---------------------|
    | Free | 3 |
    | Starter | 25 |
    | Pro | Unlimited |
    | Enterprise | Unlimited |

    Quota headers are included in quote operations:
    - `X-Quota-Limit`: Maximum quotes for period
    - `X-Quota-Used`: Quotes used this period
    - `X-Quota-Reset`: ISO timestamp when quota resets

  version: 1.0.0
  contact:
    name: QuoteMyAV Support
    email: support@quotemyav.com
  license:
    name: Proprietary
    url: https://quotemyav.com/terms
servers:
  - url: https://api.quotemyav.com/v1
    description: Production
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: List webhook endpoints
      description: Returns the user's webhook endpoints (Enterprise only)
      operationId: listWebhooks
      responses:
        '200':
          description: List of webhook endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEndpoint'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierRequired'

    post:
      tags: [Webhooks]
      summary: Create webhook endpoint
      description: |
        Creates a new webhook endpoint (Enterprise only).
        If no secret is provided, one will be generated and returned.
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookCreate'
      responses:
        '201':
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/WebhookEndpoint'
                      - type: object
                        properties:
                          secret:
                            type: string
                            description: The webhook secret (only shown once if generated)
                  warning:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierRequired'
  /webhooks/{webhookId}:
    get:
      tags: [Webhooks]
      summary: Get webhook endpoint
      description: Returns a webhook endpoint's details
      operationId: getWebhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/WebhookEndpoint'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierRequired'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Webhooks]
      summary: Update webhook endpoint
      description: Updates a webhook endpoint
      operationId: updateWebhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookUpdate'
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/WebhookEndpoint'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierRequired'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Webhooks]
      summary: Delete webhook endpoint
      description: Deletes a webhook endpoint
      operationId: deleteWebhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '204':
          description: Webhook deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierRequired'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /webhooks/{webhookId}/test:
    post:
      tags: [Webhooks]
      summary: Send test webhook
      description: Sends a test webhook to verify the endpoint is working
      operationId: testWebhook
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                      statusCode:
                        type: integer
                        nullable: true
                      message:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/TierRequired'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: QuoteMyAV API key (Pro/Enterprise only)
  parameters:
...
    WebhookId:
      name: webhookId
      in: path
      required: true
      description: Webhook endpoint ID
      schema:
        type: string
  schemas:
...
    WebhookEndpoint:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastDeliveryAt:
          type: string
          format: date-time
          nullable: true
        failureCount:
          type: integer
    WebhookEventType:
      type: string
      enum:
        - quote.created
        - quote.updated
        - quote.sent
        - quote.viewed
        - quote.accepted
        - quote.rejected
        - quote.expired
        - ai.job.completed
        - ai.job.failed
    WebhookCreate:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
          minItems: 1
        secret:
          type: string
          minLength: 16
          description: Optional custom secret (one will be generated if not provided)
    WebhookUpdate:
      type: object
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        isActive:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error context
            requestId:
              type: string
              description: Request ID for debugging
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AUTH_REQUIRED
              message: Authentication required
              requestId: req_abc123
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid request body
              details:
                field: clientEmail
                issue: Invalid email format
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: QUOTE_NOT_FOUND
              message: Quote not found
              details:
                quoteId: QM-xyz789
...
    TierRequired:
      description: Higher subscription tier required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: TIER_REQUIRED
              message: This feature requires Pro or Enterprise tier
              details:
                currentTier: starter
                requiredTiers:
                  - pro
                  - enterprise